Refactor: Split up Navigator and PageDescriptor (#826)
