+++ react/src/components/header.tsx
@@ -3,8 +3,8 @@ import React, { ReactNode, useContext, useState } from 'react'
 import '../common.css'
 import './header.css'
 import flag_dimensions from '../data/flag_dimensions'
+import { Navigator } from '../navigation/Navigator'
 import { universePath } from '../navigation/links'
-import { Navigator } from '../navigation/navigator'
 import { useColors } from '../page_template/colors'
 import { useUniverse } from '../universe'
 import { useMobileLayout } from '../utils/responsive'
