+++ react/src/components/map.tsx
@@ -5,8 +5,8 @@ import React, { ReactNode } from 'react'
 
 import { loadProtobuf } from '../load_json'
 import { Basemap } from '../mapper/settings'
+import { Navigator } from '../navigation/Navigator'
 import { shapeLink } from '../navigation/links'
-import { Navigator } from '../navigation/navigator'
 import { useColors } from '../page_template/colors'
 import { relatedSettingsKeys, relationshipKey, useSetting, useSettings } from '../page_template/settings'
 import { randomColor } from '../utils/color'
