+++ react/src/components/mapper-panel.tsx
@@ -10,8 +10,8 @@ import statNames from '../data/statistic_name_list'
 import { loadProtobuf } from '../load_json'
 import { Keypoints, Ramp, parseRamp } from '../mapper/ramps'
 import { Basemap, ColorStat, ColorStatDescriptor, FilterSettings, LineStyle, MapSettings, MapperSettings, parseColorStat } from '../mapper/settings'
+import { Navigator } from '../navigation/Navigator'
 import { consolidatedShapeLink, consolidatedStatsLink } from '../navigation/links'
-import { Navigator } from '../navigation/navigator'
 import { PageTemplate } from '../page_template/template'
 import { interpolateColor } from '../utils/color'
 import { ConsolidatedShapes, ConsolidatedStatistics, Feature, IAllStats } from '../utils/protos'
