+++ react/src/components/related-button.tsx
@@ -3,7 +3,7 @@ import React, { ReactNode, useContext, useId } from 'react'
 import './related.css'
 import type_ordering_idx from '../data/type_ordering_idx'
 import type_to_type_category from '../data/type_to_type_category'
-import { Navigator } from '../navigation/navigator'
+import { Navigator } from '../navigation/Navigator'
 import { HueColors } from '../page_template/color-themes'
 import { useColors } from '../page_template/colors'
 import { relationshipKey, useSetting } from '../page_template/settings'
