+++ react/src/components/table.tsx
@@ -3,8 +3,8 @@ import ContentEditable, { ContentEditableEvent } from 'react-contenteditable'
 
 import { loadOrdering } from '../load_json'
 import './table.css'
+import { Navigator } from '../navigation/Navigator'
 import { statisticDescriptor } from '../navigation/links'
-import { Navigator } from '../navigation/navigator'
 import { Colors } from '../page_template/color-themes'
 import { useColors } from '../page_template/colors'
 import { MobileArticlePointers, rowExpandedKey, Settings, useSetting } from '../page_template/settings'
