+++ react/src/universe.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react'
 
 import universes_default from './data/universes_default'
-import { Navigator } from './navigation/navigator'
+import { Navigator } from './navigation/Navigator'
 
 export function useUniverse(): string {
     return useContext(Navigator.Context).useUniverse() ?? (() => { throw new Error(`No universe for current page`) })()
